requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
pytz>=2023.3
python-dotenv>=1.0.0
anthropic>=0.18.0
